+++ CHANGELOG.md
@@ -8,6 +8,7 @@ All notable changes to this project will be documented in this file.
 
 - Accept scheduler default result-proxy as a web2 secret to fallback on it when no proxy
   is specified in deal parameters. (#273)
+- Configure the SMS at startup to generate Scone sessions in Hardware or MAA mode. (#275)
 
 ### Quality